+++ components/Spotify/index.ts
@@ -1,0 +1,1 @@
+export { Spotify } from './Spotify'
