+++ components/Spotify/types.ts
@@ -1,0 +1,6 @@
+export type NowPlaying = {
+  isPlaying: boolean
+  title?: string
+  artist?: string
+  songURL?: string
+}
