--- pages/_app.tsx
@@ -1,16 +1,0 @@
-import '../styles/globals.css'
-import type { AppProps } from 'next/app'
-import Header from '../components/Header'
-import Footer from '../components/Footer'
-
-function MyApp({ Component, pageProps }: AppProps) {
-  return (
-    <div className="flex flex-col items-center justify-between min-h-screen">
-      <Header />
-      <Component {...pageProps} />
-      <Footer />
-    </div>
-  )
-}
-
-export default MyApp
